+++ src/main/java/com/example/repo/TrainingTypeRepository.java
@@ -1,6 +1,5 @@
 package com.example.repo;
 
-import com.example.annotation.Storage;
 import com.example.model.TrainingType;
 import com.example.storage.StorageComponent;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +14,7 @@ public class TrainingTypeRepository implements CrudRepository<TrainingType> {
     private StorageComponent storageComponent;
 
     @Override
-    public Map<Long, TrainingType> findAll() {
+    public Map<Integer, TrainingType> findAll() {
         return storageComponent.getTrainingTypeMap();
     }
 }
