+++ src/main/java/com/example/service/CrudService.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 public interface CrudService<T> {
 
-    Map<Long, T> list() throws IOException, ParseException;
+    Map<Integer, T> list() throws IOException, ParseException;
 
     T create(T t) throws IOException, ParseException;
